feat:#304-initial structure
