+++ tailwind.config.ts
@@ -63,6 +63,7 @@ const config: Config = {
         chipServiceShadow: '4px 4px 10px 0px rgba(230, 230, 230, 0.25)',
         chipAreaShadow: '4px 4px 10px 0px rgba(230, 230, 230, 0.161)',
         profileShadow: '2px 2px 8px 0px rgba(224,224,224,0.2)',
+        MapTabShadow: '4px 4px 10px 0px rgba(230, 230, 230, 0.2)',
       },
     },
   },
